Organizar funcoes do DataManager e corrigir warnings
