--- macroChallengeApp/macroChallengeApp/Controller/DataManager.swift
@@ -1,883 +1,0 @@
-//
-//  DataManager.swift
-//  macroChallengeApp
-//
-//  Created by Beatriz Duque on 05/09/22.
-//
-
-import Foundation
-import UIKit
-import FirebaseCore
-import FirebaseAnalytics
-
-class DataManager {
-    public static var shared = DataManager()
-    let baseURL: String = "https://macrotrip-dev.herokuapp.com/"
-    
-    public let imageCash = NSCache<NSNumber, UIImage>()
-    
-    var roadmaps: Roadmaps?
-    var user: User?
-    var activity: Activity?
-    var day: Day?
-    var like: Like?
-    var delegate: UserLoggedInDelegate?
-    
-    // MARK: - Load Data
-    public func loadData(_ completion: @escaping (() -> Void), dataURL: String, dataType: DataType) {
-        let session = URLSession.shared
-        let url = URL(string: baseURL + dataURL)!
-        
-        // resposta vem no data (json), error vem no error
-        let task = session.dataTask(with: url) { data, _, error in
-            guard let data = data else { return }
-            if error != nil {
-                print(String(describing: error?.localizedDescription))
-            }
-            self.getDataType(dataType: dataType, data: data)
-        }
-        task.resume()
-    }
-    
-    private func getDataType(dataType: DataType, data: Data ) {
-        switch dataType {
-        case .USER:
-            do {
-                self.user = try JSONDecoder().decode(User.self, from: data)
-            } catch {
-                print("Parse Error")
-            }
-        case .ROADMAPS:
-            do {
-                self.roadmaps = try JSONDecoder().decode(Roadmaps.self, from: data)
-            } catch {
-                print("Parse Error")
-            }
-        case .ACTIVITY:
-            do {
-                self.activity = try JSONDecoder().decode(Activity.self, from: data)
-            } catch {
-                print("Parse Error")
-            }
-        case .DAY:
-            do {
-                self.day = try JSONDecoder().decode(Day.self, from: data)
-            } catch {
-                print("Parse Error")
-            }
-        case .LIKE:
-            do {
-                self.like = try JSONDecoder().decode(Like.self, from: data)
-            } catch {
-                print("Parse Error")
-            }
-        }
-    }
-    
-    func postUser(username: String, usernameApp: String, name: String, photoId: String, password: String, _ completion: @escaping (() -> Void)) {
-        let user: [String: Any] = [
-            "username": username,
-            "usernameApp": usernameApp,
-            "name": name,
-            "photoId": "icon",
-            "password": password
-        ]
-        
-        let session = URLSession.shared
-        guard let url = URL(string: baseURL + "users") else { return }
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        
-        do {
-            request.httpBody = try JSONSerialization.data(withJSONObject: user, options: .prettyPrinted)
-        } catch {
-            print(error.localizedDescription)
-        }
-        
-        request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.addValue("application/json", forHTTPHeaderField: "Accept")
-        
-        let task = session.dataTask(with: request) { data, response, error in
-            if let error = error {
-                print(error)
-            } else if data != nil {
-                if let httpResponse = response as? HTTPURLResponse {
-                    DispatchQueue.main.async {
-                        completion()
-                    }
-                    if httpResponse.statusCode == 200 {
-                        print("Criou")
-                        FirebaseManager.shared.createAnalyticsEvent(event: AnalyticsEventSignUp)
-
-                    }
-                    
-                }
-            } else {
-                // Handle unexpected error
-            }
-        }
-        task.resume()
-    }
-    
-    func postLogin(username: String, password: String) {
-        let user: [String: Any] = [
-            "username": username,
-            "password": password
-        ]
-        
-        let session = URLSession.shared
-        guard let url = URL(string: baseURL + "login") else { return }
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        
-        do {
-            request.httpBody = try JSONSerialization.data(withJSONObject: user, options: .prettyPrinted)
-        } catch {
-            print(error.localizedDescription)
-        }
-        
-        request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.addValue("application/json", forHTTPHeaderField: "Accept")
-        
-        let task = session.dataTask(with: request) { data, response, error in
-            if let error = error {
-                print(error)
-            } else if data != nil {
-                if let httpResponse = response as? HTTPURLResponse {
-                    if httpResponse.statusCode == 200 {
-                        do {
-                            try KeychainManager.shared.save(username, service: "username", account: "explorer")
-                        } catch {
-                            print(error)
-                        }
-                        
-                        let jwtToken = httpResponse.value(forHTTPHeaderField: "Authorization")
-                        UserDefaults.standard.setValue(jwtToken, forKey: "authorization")
-                        UserDefaults.standard.set(true, forKey: "isUserLoggedIn")
-                        
-                        FirebaseManager.shared.createAnalyticsEvent(event: AnalyticsEventLogin)
-
-                        
-                        DispatchQueue.main.async {
-                            self.delegate?.finishLogin()
-                        }
-                    }
-                }
-            } else {
-                // Handle unexpected error
-            }
-        }
-        task.resume()
-    }
-    
-    func postRoadmap(roadmap: Roadmaps, roadmapCore: RoadmapLocal, daysCore: [DayLocal], selectedImage: UIImage? = nil) {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "d/M/y"
-        
-        let keyWords = ["trip", "viagem", "travel", "viaggiare"]
-        let codeTrip = "\(Int.random(in: 0..<1_000_000))\(keyWords.randomElement() ?? "space")\(Int.random(in: 0..<1_000_000))"
-        
-        let roadmap: [String: Any] = [
-            "name": roadmap.name,
-            "location": roadmap.location,
-            "budget": 0,
-            "dayCount": roadmap.dayCount,
-            "dateInitial": roadmap.dateInitial,
-            "dateFinal": roadmap.dateFinal,
-            "peopleCount": roadmap.peopleCount,
-            "imageId": setupImage(category: roadmap.category),
-            "category": roadmap.category,
-            "isShared": roadmap.isShared,
-            "isPublic": roadmap.isPublic,
-            "shareKey": codeTrip,
-            "createdAt": dateFormatter.string(from: Date()),
-            "currency": roadmap.currency,
-            "likesCount": 0
-        ]
-        
-        let session = URLSession.shared
-        if let data = KeychainManager.shared.read(service: "username", account: "explorer") {
-            let userID = String(data: data, encoding: .utf8)!
-            guard let url = URL(string: baseURL + "roadmaps/users/\(userID)") else { return }
-            var request = URLRequest(url: url)
-            request.httpMethod = "POST"
-            if let token = UserDefaults.standard.string(forKey: "authorization") {
-                request.setValue(token, forHTTPHeaderField: "Authorization")
-                
-                do {
-                    request.httpBody = try JSONSerialization.data(withJSONObject: roadmap, options: .prettyPrinted)
-                } catch {
-                    print(error.localizedDescription)
-                }
-                
-                request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-                request.addValue("application/json", forHTTPHeaderField: "Accept")
-                
-                let task = session.dataTask(with: request) { data, response, error in
-                    guard let data = data else { return }
-                    if error != nil {
-                        print(String(describing: error?.localizedDescription))
-                    }
-                    if let httpResponse = response as? HTTPURLResponse {
-                        if httpResponse.statusCode == 200 {
-                            do {
-                                // tentar transformar os dados no tipo Cohort
-                                let roadmapResponse = try JSONDecoder().decode(RoadmapDTO.self, from: data)
-                                
-                                self.postDays(roadmapId: roadmapResponse.id, daysCore: daysCore)
-                                
-                                roadmapCore.id = Int32(roadmapResponse.id)
-                                roadmapCore.shareKey = codeTrip
-                                RoadmapRepository.shared.saveContext()
-                                
-                                if let selectedImage = selectedImage {
-                                    FirebaseManager.shared.uploadImageRoadmap(image: selectedImage, roadmapId: roadmapResponse.id, roadmapCore: roadmapCore)
-                                }
-                            } catch {
-                                print(error)
-                            }
-                        }
-                    }
-                }
-                task.resume()
-            }
-        }
-    }
-    
-    func getUser(username: String, _ completion: @escaping ((_ user: User) -> Void)) {
-        var user: User?
-        let session: URLSession = URLSession.shared
-        let url: URL = URL(string: baseURL + "users/\(username)")!
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.addValue("application/json", forHTTPHeaderField: "Accept")
-        
-        if let token = UserDefaults.standard.string(forKey: "authorization") {
-            request.setValue(token, forHTTPHeaderField: "Authorization")
-            let task = session.dataTask(with: request) { data, response, error in
-                guard let data = data else { return }
-                if error != nil {
-                    print(String(describing: error?.localizedDescription))
-                }
-                
-                do {
-                    // tentar transformar os dados no tipo Cohort
-                    user = try JSONDecoder().decode(User.self, from: data)
-                    
-                    DispatchQueue.main.async {
-                        completion(user!)
-                    }
-                } catch {
-                    print(error)
-                }
-            }
-            task.resume()
-            
-        }
-    }
-    
-    func getPublicRoadmaps(_ completion: @escaping ((_ roadmaps: [RoadmapDTO]) -> Void)) {
-        var roadmaps: [RoadmapDTO] = []
-        let session: URLSession = URLSession.shared
-        let url: URL = URL(string: baseURL + "roadmaps")!
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.addValue("application/json", forHTTPHeaderField: "Accept")
-        
-        let task = session.dataTask(with: request) { data, _, error in
-            guard let data = data else { return }
-            if error != nil {
-                print(String(describing: error?.localizedDescription))
-            }
-            
-            do {
-                roadmaps = try JSONDecoder().decode([RoadmapDTO].self, from: data)
-                DispatchQueue.main.async {
-                    completion(roadmaps)
-                }
-            } catch {
-                print("DEU RUIM NO PARSE")
-            }
-        }
-        task.resume()
-    }
-    
-    func getLikedRoadmaps(_ completion: @escaping ((_ roadmaps: [RoadmapDTO]) -> Void)) {
-        var roadmaps: [RoadmapDTO] = []
-        let session: URLSession = URLSession.shared
-        
-        if let data = KeychainManager.shared.read(service: "username", account: "explorer") {
-            let userID = String(data: data, encoding: .utf8)!
-            let url: URL = URL(string: baseURL + "likes/users/\(userID)")!
-            var request = URLRequest(url: url)
-            request.httpMethod = "GET"
-            request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-            request.addValue("application/json", forHTTPHeaderField: "Accept")
-            
-            if let token = UserDefaults.standard.string(forKey: "authorization") {
-                request.setValue(token, forHTTPHeaderField: "Authorization")
-                let task = session.dataTask(with: request) { data, _, error in
-                    guard let data = data else { return }
-                    if error != nil {
-                        print(String(describing: error?.localizedDescription))
-                    }
-                    
-                    do {
-                        roadmaps = try JSONDecoder().decode([RoadmapDTO].self, from: data)
-                        
-                        DispatchQueue.main.async {
-                            completion(roadmaps)
-                        }
-                    } catch {
-                        print("DEU RUIM NO PARSE")
-                    }
-                }
-                task.resume()
-            }
-        }
-    }
-    
-    func putUser(userObj: UserLocal) {
-        let user: [String: Any] = [
-            "usernameApp": userObj.usernameApp ?? "newuser",
-            "name": userObj.name ?? "newuser",
-            "photoId": "icon"
-        ]
-        
-        if let data = KeychainManager.shared.read(service: "username", account: "explorer") {
-            let userID = String(data: data, encoding: .utf8)!
-            
-            let session = URLSession.shared
-            guard let url = URL(string: baseURL + "users/\(userID)") else { return }
-            
-            var request = URLRequest(url: url)
-            request.httpMethod = "PUT"
-            
-            do {
-                request.httpBody = try JSONSerialization.data(withJSONObject: user, options: .prettyPrinted)
-            } catch {
-                print(error.localizedDescription)
-            }
-            
-            request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-            request.addValue("application/json", forHTTPHeaderField: "Accept")
-            
-            if let token = UserDefaults.standard.string(forKey: "authorization") {
-                request.setValue(token, forHTTPHeaderField: "Authorization")
-                let task = session.dataTask(with: request) { data, _, error in
-                    if let error = error {
-                        print(error)
-                    } else if data != nil {
-                    } else {
-                        // Handle unexpected error
-                    }
-                }
-                task.resume()
-            }
-        }
-    }
-    
-    func getRoadmapById(roadmapId: Int, _ completion: @escaping ((_ roadmap: Roadmaps?) -> Void)) {
-        var roadmap: Roadmaps?
-        let session: URLSession = URLSession.shared
-        let url: URL = URL(string: baseURL + "roadmaps/\(roadmapId)")!
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.addValue("application/json", forHTTPHeaderField: "Accept")
-        
-        let task = session.dataTask(with: request) { data, _, error in
-            guard let data = data else { return }
-            if error != nil {
-                print(String(describing: error?.localizedDescription))
-            }
-            
-            do {
-                roadmap = try JSONDecoder().decode(Roadmaps.self, from: data)
-            } catch {
-                print(error)
-                print("DEU RUIM NO PARSE")
-            }
-            
-            DispatchQueue.main.async {
-                completion(roadmap)
-            }
-        }
-        task.resume()
-        
-    }
-    
-    func putRoadmap(roadmap: Roadmaps, roadmapId: Int, newDaysCore: [DayLocal], selectedImage: UIImage? = nil) {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "d/M/y"
-        
-        let roadmapJson: [String: Any] = [
-            "name": roadmap.name,
-            "location": roadmap.location,
-            "budget": 0,
-            "dayCount": roadmap.dayCount,
-            "dateInitial": roadmap.dateInitial,
-            "dateFinal": roadmap.dateFinal,
-            "peopleCount": roadmap.peopleCount,
-            "imageId": setupImage(category: roadmap.category),
-            "category": roadmap.category,
-            "isShared": roadmap.isShared,
-            "isPublic": roadmap.isPublic,
-            "shareKey": "ABC123",
-            "createdAt": dateFormatter.string(from: Date()),
-            "currency": roadmap.currency,
-            "likesCount": roadmap.likesCount
-        ]
-        
-        let session = URLSession.shared
-        guard let url = URL(string: baseURL + "roadmaps/\(roadmapId)") else { return }
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        if let token = UserDefaults.standard.string(forKey: "authorization") {
-            request.setValue(token, forHTTPHeaderField: "Authorization")
-            
-            do {
-                request.httpBody = try JSONSerialization.data(withJSONObject: roadmapJson, options: .prettyPrinted)
-            } catch {
-                print(error.localizedDescription)
-            }
-            
-            request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-            request.addValue("application/json", forHTTPHeaderField: "Accept")
-            
-            let task = session.dataTask(with: request) { data, response, error in
-                guard let data = data else { return }
-                if error != nil {
-                    print(String(describing: error?.localizedDescription))
-                }
-                if let httpResponse = response as? HTTPURLResponse {
-                    if httpResponse.statusCode == 200 {
-                        self.postDays(roadmapId: roadmapId, daysCore: newDaysCore)
-                        
-                        if let selectedImage = selectedImage {
-                            FirebaseManager.shared.uploadImageRoadmap(image: selectedImage, roadmapId: roadmapId, uuid: roadmap.imageId)
-                        }
-                        print("Atualizou")
-                    }
-                }
-            }
-            task.resume()
-        }
-    }
-    
-    func putImageRoadmap(roadmapId: Int, uuid: String) {
-        let session = URLSession.shared
-        var urlFinal = ""
-        var oldId = UserRepository.shared.getUser()[0].photoId
-        
-        if roadmapId != 0 {
-            urlFinal = "roadmaps/\(roadmapId)/\(uuid)"
-        } else {
-            if let data = KeychainManager.shared.read(service: "username", account: "explorer") {
-                let userID = String(data: data, encoding: .utf8)!
-                print(userID)
-                urlFinal = "users/\(userID)/\(uuid)"
-            }
-        }
-        
-        guard let url = URL(string: baseURL + urlFinal) else { return }
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        if let token = UserDefaults.standard.string(forKey: "authorization") {
-            request.setValue(token, forHTTPHeaderField: "Authorization")
-            
-            request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-            request.addValue("application/json", forHTTPHeaderField: "Accept")
-            
-            let task = session.dataTask(with: request) { data, response, error in
-                guard let data = data else { return }
-                if error != nil {
-                    print(String(describing: error?.localizedDescription))
-                }
-                if let httpResponse = response as? HTTPURLResponse {
-                    if httpResponse.statusCode == 200 {
-                        print("Atualizou")
-                    }
-                }
-            }
-            task.resume()
-        }
-    }
-    
-    func postDays(roadmapId: Int, daysCore: [DayLocal]) {
-        let session = URLSession.shared
-        guard let url = URL(string: baseURL + "roadmaps/\(roadmapId)/days") else { return }
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        if let token = UserDefaults.standard.string(forKey: "authorization") {
-            request.setValue(token, forHTTPHeaderField: "Authorization")
-            request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-            request.addValue("application/json", forHTTPHeaderField: "Accept")
-            
-            let task = session.dataTask(with: request) { data, response, error in
-                print(response)
-                guard let data = data else { return }
-                if error != nil {
-                    print(String(describing: error?.localizedDescription))
-                }
-                if let httpResponse = response as? HTTPURLResponse {
-                    if httpResponse.statusCode == 200 {
-                        do {
-                            // tentar transformar os dados no tipo Cohort
-                            var daysResponse = try JSONDecoder().decode([DayDTO].self, from: data)
-                            
-                            daysResponse.sort { $0.id < $1.id }
-                            
-                            for index in 0..<daysResponse.count {
-                                daysCore[index].id = Int32(daysResponse[index].id)
-                                DayRepository.shared.saveContext()
-                            }
-                            
-                        } catch {
-                            print(error)
-                        }
-                    }
-                }
-            }
-            task.resume()
-        }
-    }
-    
-    func postActivity(activity: Activity, dayId: Int, activityCore: ActivityLocal) {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "d/M/y"
-        
-        let activity: [String: Any] = [
-            "name": activity.name,
-            "tips": activity.tips,
-            "category": activity.category,
-            "location": activity.location,
-            "hour": activity.hour,
-            "budget": activity.budget,
-            "currency": activity.currency,
-            "address": activity.address,
-            "link": activity.link
-        ]
-        
-        let session = URLSession.shared
-        guard let url = URL(string: baseURL + "days/\(dayId)/activities") else { return }
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        if let token = UserDefaults.standard.string(forKey: "authorization") {
-            request.setValue(token, forHTTPHeaderField: "Authorization")
-            
-            do {
-                request.httpBody = try JSONSerialization.data(withJSONObject: activity, options: .prettyPrinted)
-            } catch {
-                print(error.localizedDescription)
-            }
-            
-            request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-            request.addValue("application/json", forHTTPHeaderField: "Accept")
-            
-            let task = session.dataTask(with: request) { data, response, error in
-                print(response)
-                guard let data = data else { return }
-                if error != nil {
-                    print(String(describing: error?.localizedDescription))
-                }
-                if let httpResponse = response as? HTTPURLResponse {
-                    if httpResponse.statusCode == 200 {
-                        do {
-                            let activityResponse = try JSONDecoder().decode(Activity.self, from: data)
-                            
-                            activityCore.id = Int32(activityResponse.id)
-                            ActivityRepository.shared.saveContext()
-                        } catch {
-                            print(error)
-                        }
-                    }
-                    
-                }
-            }
-            task.resume()
-        }
-    }
-    
-    func putActivity(activity: Activity, dayId: Int) {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "d/M/y"
-        
-        let activityNew: [String: Any] = [
-            "name": activity.name,
-            "tips": activity.tips,
-            "category": activity.category,
-            "location": activity.location,
-            "hour": activity.hour,
-            "budget": activity.budget,
-            "currency": activity.currency,
-            "address": activity.address,
-            "link": activity.link
-        ]
-        
-        let session = URLSession.shared
-        
-        guard let url = URL(string: baseURL + "days/\(dayId)/activities/\(activity.id)") else { return }
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        if let token = UserDefaults.standard.string(forKey: "authorization") {
-            request.setValue(token, forHTTPHeaderField: "Authorization")
-            
-            do {
-                request.httpBody = try JSONSerialization.data(withJSONObject: activityNew, options: .prettyPrinted)
-            } catch {
-                print(error.localizedDescription)
-            }
-            
-            request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-            request.addValue("application/json", forHTTPHeaderField: "Accept")
-            
-            let task = session.dataTask(with: request) { data, response, error in
-                print(response)
-                guard let data = data else { return }
-                if error != nil {
-                    print(String(describing: error?.localizedDescription))
-                }
-                if let httpResponse = response as? HTTPURLResponse {
-                    if httpResponse.statusCode == 200 {
-                        do {
-                        } catch {
-                            print(error)
-                        }
-                    }
-                }
-            }
-            task.resume()
-        }
-    }
-    
-    func deleteObjectBack(objectID: Int = 0, username: String = "", urlPrefix: String) {
-        let session: URLSession = URLSession.shared
-        var url: URL
-        if username.isEmpty {
-            url = URL(string: baseURL + "\(urlPrefix)/\(objectID)")!
-        } else {
-            url = URL(string: baseURL + "\(urlPrefix)/\(username)")!
-        }
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "DELETE"
-        request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.addValue("application/json", forHTTPHeaderField: "Accept")
-        
-        if let token = UserDefaults.standard.string(forKey: "authorization") {
-            request.setValue(token, forHTTPHeaderField: "Authorization")
-            let task = session.dataTask(with: request) { data, _, error in
-                guard data != nil else { return }
-                if error != nil {
-                    print(String(describing: error?.localizedDescription))
-                }
-            }
-            task.resume()
-        }
-    }
-    
-    func joinRoadmap(roadmapKey: String) {
-        let session = URLSession.shared
-        if let data = KeychainManager.shared.read(service: "username", account: "explorer") {
-            let userID = String(data: data, encoding: .utf8)!
-            guard let url = URL(string: baseURL + "roadmaps/\(roadmapKey)/users/\(userID)") else { return }
-            var request = URLRequest(url: url)
-            request.httpMethod = "POST"
-            if let token = UserDefaults.standard.string(forKey: "authorization") {
-                request.setValue(token, forHTTPHeaderField: "Authorization")
-                
-                request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-                request.addValue("application/json", forHTTPHeaderField: "Accept")
-                
-                let task = session.dataTask(with: request) { data, response, error in
-                    print(response ?? "sem resposta")
-                    guard let data = data else { return }
-                    if error != nil {
-                        print(String(describing: error?.localizedDescription))
-                    }
-                    if let httpResponse = response as? HTTPURLResponse {
-                        if httpResponse.statusCode == 200 {
-                            do {
-                                let roadmapJoin = try JSONDecoder().decode(Roadmaps.self, from: data)
-                                let newRoadmap = RoadmapRepository.shared.createRoadmap(roadmap: roadmapJoin, isNew: false)
-                                newRoadmap.id = Int32(roadmapJoin.id)
-                                newRoadmap.shareKey = roadmapJoin.shareKey
-                                newRoadmap.isShared = true
-                                
-                                var days = newRoadmap.day?.allObjects as [DayLocal]
-                                var roadmapDays = roadmapJoin.days
-                                
-                                days.sort { $0.id < $1.id }
-                                roadmapDays.sort { $0.id < $1.id }
-                                
-                                for index in 0..<roadmapDays.count {
-                                    days[index].id = Int32(roadmapDays[index].id)
-                                    let activiyArray = roadmapDays[index].activity
-                                    for activity in activiyArray {
-                                        _ = ActivityRepository.shared.createActivity(day: days[index], activity: activity, isNew: false)
-                                    }
-                                }
-                                RoadmapRepository.shared.saveContext()
-                                DayRepository.shared.saveContext()
-                            } catch {
-                                print(error)
-                            }
-                        }
-                    }
-                }
-                task.resume()
-            }
-        }
-    }
-    
-    func getLike(roadmapId: Int, _ completion: @escaping ((_ response: Int) -> Void)) {
-        let session: URLSession = URLSession.shared
-        
-        if let data = KeychainManager.shared.read(service: "username", account: "explorer") {
-            let userID = String(data: data, encoding: .utf8)!
-            print(userID)
-            let url: URL = URL(string: baseURL + "likes/users/\(userID)/roadmaps/\(roadmapId)")!
-            
-            var request = URLRequest(url: url)
-            request.httpMethod = "GET"
-            request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-            request.addValue("application/json", forHTTPHeaderField: "Accept")
-            
-            if let token = UserDefaults.standard.string(forKey: "authorization") {
-                request.setValue(token, forHTTPHeaderField: "Authorization")
-                let task = session.dataTask(with: request) { data, _, error in
-                    guard let data = data else { return }
-                    if error != nil {
-                        print(String(describing: error?.localizedDescription))
-                    }
-                    
-                    do {
-                        let stringInt = String(data: data, encoding: String.Encoding.utf8)
-                        var likeId: Int = 0
-                        if let likeIdConvert = Int(stringInt ?? "0") {
-                            likeId = likeIdConvert
-                        }
-                        DispatchQueue.main.async {
-                            completion(likeId)
-                        }
-                    } catch {
-                        print(error)
-                        print("DEU RUIM NO PARSE")
-                    }
-                }
-                task.resume()
-            }
-        }
-    }
-    
-    func postLike(roadmapId: Int, _ completion: @escaping ((_ response: Int) -> Void)) {
-        let session: URLSession = URLSession.shared
-        
-        if let data = KeychainManager.shared.read(service: "username", account: "explorer") {
-            let userID = String(data: data, encoding: .utf8)!
-            let url: URL = URL(string: baseURL + "users/\(userID)/roadmaps/\(roadmapId)/likes")!
-            
-            var request = URLRequest(url: url)
-            request.httpMethod = "POST"
-            request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-            request.addValue("application/json", forHTTPHeaderField: "Accept")
-            
-            if let token = UserDefaults.standard.string(forKey: "authorization") {
-                request.setValue(token, forHTTPHeaderField: "Authorization")
-                let task = session.dataTask(with: request) { data, response, error in
-                    print(response)
-                    guard let data = data else { return }
-                    if error != nil {
-                        print(String(describing: error?.localizedDescription))
-                    }
-                    
-                    do {
-                        let stringInt = String(data: data, encoding: String.Encoding.utf8)
-                        var likeId: Int = 0
-                        if let likeIdConvert = Int(stringInt ?? "0") {
-                            likeId = likeIdConvert
-                        }
-                        DispatchQueue.main.async {
-                            completion(likeId)
-                        }
-                    } catch {
-                        print(error)
-                        print("DEU RUIM NO PARSE")
-                    }
-                }
-                task.resume()
-            }
-            
-        }
-    }
-    
-    func getRoadmapUserImage(roadmapId: Int, _ completion: @escaping ((_ uuidUser: String, _ username: String) -> Void)) {
-        let session: URLSession = URLSession.shared
-        let url: URL = URL(string: baseURL + "roadmaps/\(roadmapId)/userImage")!
-        
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.addValue("application/json", forHTTPHeaderField: "Accept")
-        
-        if let token = UserDefaults.standard.string(forKey: "authorization") {
-            request.setValue(token, forHTTPHeaderField: "Authorization")
-            let task = session.dataTask(with: request) { data, response, error in
-                print(response)
-                guard let data = data else { return }
-                if error != nil {
-                    print(String(describing: error?.localizedDescription))
-                }
-                
-                do {
-                    // tentar transformar os dados no tipo Cohort
-                    let result = try JSONDecoder().decode([String].self, from: data)
-                    DispatchQueue.main.async {
-                        completion(result[1], result[0])
-                    }
-                    
-                } catch {
-                    print(error)
-                }
-            }
-            task.resume()
-            
-        }
-    }
-    
-    func decodeType<T: Codable>(_ class: T, data: Data) -> T? {
-        do {
-            let newData = try JSONDecoder().decode(T.self, from: data)
-            return newData
-        } catch {
-            print("Parse Error")
-        }
-        return nil
-    }
-    
-    func setupImage(category: String) -> String {
-        return String("\(category)Cover")
-    }
-}
-
-struct FailableDecodable<Base: Decodable>: Decodable {
-    let base: Base?
-    
-    init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-        self.base = try? container.decode(Base.self)
-    }
-}
